feat: server-render the request page so crawlers get the request
The previous commits stopped /request/<id> returning 500, but the page
still served nothing useful: it fetched on the client, so the HTML was a
skeleton with zero request data. An SEO audit flagged these URLs, and a
200 with an empty shell is barely better than the 500 it replaced.

page.tsx is now an async server component that awaits fetchRequest and
passes the Channel to a new client component, request-details.tsx, which
keeps everything genuinely browser-bound: PDF export, the JSON viewer,
the clipboard buttons, TimeAgo, useState. The relocation is verbatim —
same markup, classNames, copy and table structure.

Verified against a local stand-in backend: the served HTML now contains
the gateway, payee, payer, expected amount, payment reference, encryption
status and the transactions table. Before, none of it was there.

Two things this uncovered, both worth reading:

**json-edit-react cannot be server-rendered.** Its theme provider calls
`document.documentElement.style.setProperty("--jer-highlight-color", …)`
during render. Once the details actually server-rendered, that threw
`ReferenceError: document is not defined`, React bailed out to client
rendering, and the HTML went back to being an empty shell — the original
bug's twin, hidden until now because SSR had never got as far as
rendering this component. It is loaded via next/dynamic with ssr:false.
The raw JSON blob is not the indexable content, so nothing is lost.
Note this is a *render-time* touch of document, not an import-time one,
which is why an import-level audit of the same file came back clean.

**A missing request still returns 200, not 404.** It renders the
not-found UI, but the status is wrong, and this is a framework limit
rather than something left undone. The route streams
(Transfer-Encoding: chunked), so the status commits when the shell
flushes, before the lookup resolves; notFound() can then only swap the
UI. Tried and measured: notFound() from the page body → 200, and
notFound() from generateMetadata → also 200. A real 404 needs the
existence check to run before the response commits, which means
middleware and a Hasura call on every request — a bigger trade than this
commit should make unilaterally.

So the actual harm is addressed instead: a missing request is served
`<meta name="robots" content="noindex, nofollow">`, which is what stops
soft 404s accumulating in the search index. Valid requests are not
noindexed — checked both ways.

Also: generateMetadata gives these pages a real title and description
for search results and link previews, and the request lookup is wrapped
in React's cache() keyed on the id so generateMetadata and the page share
one network call. The fetch-level cache cannot do that here — graphql-
request issues a POST, which Next's Data Cache does not dedupe, and the
cache() inside graphQlClient is keyed on a fresh RequestInit per call.

The two dependent queries (payments, SRF deployments) stay client-side:
they hang off a payment reference derived from the request, they are the
polling surface via refetchInterval, and keeping them there guarantees
neither can affect the HTTP status or block the request's own content.
The four cells they feed — Status, Balance, Modified, and the payments
table — render their own inline loading state, so they no longer briefly
display wrong values (Balance 0, Status from the last transaction) before
the query lands.
